+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - Add support for the Steam Deck controller.
 - Add 'pacman-static' for more stable upgrades.
 - Change SteamOS packages to be optionally installed as part of the first-time setup.
+- Add support for the Vulkan graphics pipeline library on Intel.
 
 ## 3.2.1 - The SteamOS 3.4 Update